Abstracted imports to their respective child classes, improved exception output and modified tests accordingly
a
