+++ python/TestHarness/testers/XMLDiff.py
@@ -15,7 +15,7 @@ class XMLDiff(SchemaDiff):
     def validParams():
         params = SchemaDiff.validParams()
         params.addRequiredParam('xmldiff',   [], "A list of XML files to compare.")
-        params.addParam('ignored_attributes',  [], "Deprecated. Items in the JSON that the differ will ignore. This is functionally identical to ignored_items inside of SchemaDiff.")
+        params.addParam('ignored_attributes',  [], "Deprecated. Items in the XML that the differ will ignore. This is functionally identical to ignored_items inside of SchemaDiff.")
         return params
 
     def __init__(self, name, params):
@@ -33,4 +33,22 @@ class XMLDiff(SchemaDiff):
         return SchemaDiff.processResults(self, moose_dir, options, output)
 
     def load_file(self, path1):
-        return (self.import_xml(path1), 1)
+        import xmltodict
+        with open(path1,"r") as f:
+            xml = xmltodict.parse(f.read())
+            return self.fix_xml_arrays(xml)
+
+    #Many xml files in MOOSE don't actually store their data as a proper array, but as a giant string.
+    #This should fix the giant strings as lists split by whitespace, in-place in the dict.
+    def fix_xml_arrays(self, data):
+        for k, v in data.copy().items():
+            if isinstance(v, dict):
+                data[k] = self.fix_xml_arrays(v)
+            elif isinstance(v, list):
+                data[k] = [self.fix_xml_arrays(i) for i in v]
+            elif k == '#text':
+                del data[k]
+                data[k] = v.split()
+        return data
+
+
